refactor: move authorization code oauth to own dir - NoWorkspaceEntry - WorkspaceEntry - OAuthWindow - fetchOAuthPopupURL
