+++ src/components/auth/Oauth/AuthorizationCode/NoWorkspaceEntry/LandingContentProps.tsx
