+++ src/components/auth/Oauth/AuthorizationCode/NoWorkspaceEntry/NoWorkspaceEntryContent.tsx
