+++ src/components/auth/Oauth/AuthorizationCode/OAuthWindow/windowHelpers.tsx
