+++ src/components/auth/Oauth/AuthorizationCode/WorkspaceEntry/WorkspaceEntryProps.tsx
